# docs/api/openapi.json
{
  "components": {
    "responses": {
      "Error": {
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "error": {
                  "type": "string"
                }
              },
              "required": [
                "error"
              ],
              "type": "object"
            }
          },
          "text/plain": {
            "schema": {
              "type": "string"
            }
          }
        },
        "description": "400 invalid request; 401 missing/incorrect configured API key or invalid/expired session; 403 forbidden or unverified LAN identity when no API key is configured; 404 absent resource; 405 unsupported method; 409 CA changed; 413 body over 1 KiB; 415 wrong media type; 500 persistence failure; 503 starting/reloading. Route/method errors and missing certificates may be plain text."
      }
    },
    "schemas": {
...
      "Certificate": {
        "properties": {
          "fingerprint": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "test_available": {
            "type": "boolean"
          },
          "test_generation": {
            "type": "string",
            "description": "Current diagnostic generation. Discard browser test results when this changes."
          },
          "test_mitm_origins": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Exact virtual HTTPS origins allowed for MITM challenges. CA challenges use HTTPS on the current API socket."
          }
        },
        "required": [
          "name",
          "fingerprint",
          "test_available"
        ],
        "type": "object"
      },
      "CertificateTest": {
        "type": "object",
        "description": "A one-minute browser challenge bound to the verified caller and current configuration. Observations are not a device-wide CA trust verdict; the browser must receive and validate the matching proof over HTTPS with normal certificate verification.",
        "properties": {
          "id": { "type": "string" },
          "ca_fingerprint": { "type": "string" },
          "started_at": { "type": "string", "format": "date-time" },
          "expires_at": { "type": "string", "format": "date-time" },
          "trust_url": { "type": "string" },
          "mitm_url": { "type": "string" },
          "mitm_enabled": { "type": "boolean" },
          "trust_observed": { "type": "boolean" },
          "mitm_observed": { "type": "boolean" }
        },
        "required": ["id", "ca_fingerprint", "started_at", "expires_at", "trust_url", "mitm_url", "mitm_enabled", "trust_observed", "mitm_observed"]
      },
...
      "DeviceStatus": {
        "type": "object",
        "description": "The requesting device's MAC-attributed upstream paths. IPv4 and IPv6 share the MAC identity. Counters survive reload and reset on daemon restart. Kernel passthrough, local HTTP/DNS responses and daemon-originated work are excluded; splice deltas are included.",
        "properties": {
          "device": { "$ref": "#/components/schemas/DeviceState" },
          "scope": { "type": "string", "enum": ["userspace_upstream"] },
          "started_at": { "type": "string", "format": "date-time" },
          "stats": { "$ref": "#/components/schemas/PathStats" },
          "networks": {
            "type": "array", "minItems": 4, "maxItems": 4,
            "items": { "$ref": "#/components/schemas/PathStats" }
          },
          "direct_fallback_connections": { "type": "integer", "format": "int64" },
          "outbounds": { "type": "array", "items": { "$ref": "#/components/schemas/DeviceOutboundStatus" } }
        },
        "required": ["device", "scope", "started_at", "stats", "networks", "direct_fallback_connections", "outbounds"]
      },
...
      "ClientSetState": {
        "properties": {
          "description": {
            "type": "string"
          },
          "joined": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "description",
          "joined"
        ],
        "type": "object"
      },
      "DeviceState": {
        "properties": {
          "mac": {
            "type": "string"
          },
          "mitm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MITMState"
              },
              {
                "type": "null"
              }
            ]
          },
          "sets": {
            "items": {
              "$ref": "#/components/schemas/ClientSetState"
            },
            "type": "array"
          },
          "source_ip": {
            "type": "string"
          }
        },
        "required": [
          "source_ip",
          "mac",
          "sets"
        ],
        "type": "object"
      },
...
      "MITMState": {
        "properties": {
          "ca_fingerprint": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "override": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "enabled",
          "override",
          "ca_fingerprint"
        ],
        "type": "object"
      },
      "ResourceRefreshStatus": {
        "type": "object",
        "properties": {
          "run": { "type": "integer" },
          "state": { "type": "string", "enum": ["idle", "queued", "running", "completed", "failed"] },
          "trigger": { "type": "string", "enum": ["api", "automatic"] },
          "started_at": { "type": "string", "format": "date-time" },
          "finished_at": { "type": "string", "format": "date-time" },
          "next_check": { "type": "string", "format": "date-time" },
          "result": { "type": "string", "description": "Publication result, including the number of groups using previous contents after a failed refresh." },
          "error": { "type": "string" }
        },
        "required": ["run", "state"]
      },
      "ScriptRunRequest": {
        "type": "object",
        "properties": {
          "module": { "type": "string", "description": "Required when the script name is not unique within the instance." },
          "script": { "type": "string", "minLength": 1 }
        },
        "required": ["script"],
        "additionalProperties": false
      },
      "ScriptRunResponse": {
        "type": "object",
        "properties": {
          "instance": { "type": "string" },
          "module": { "type": "string" },
          "run": { "type": "integer", "minimum": 1 },
          "status": { "$ref": "#/components/schemas/ScriptTaskStatus" }
        },
        "required": ["instance", "module", "run", "status"]
      },
...
      "PathStats": {
        "properties": {
          "active_connections": {
            "format": "int64",
            "type": "integer"
          },
          "download_bytes": {
            "description": "Unsigned 64-bit count; JavaScript clients must account for integer precision above 2^53-1.",
            "minimum": 0,
            "type": "integer"
          },
          "history": {
            "$ref": "#/components/schemas/TrafficHistory"
          },
          "total_connections": {
            "format": "int64",
            "type": "integer"
          },
          "upload_bytes": {
            "description": "Unsigned 64-bit count; JavaScript clients must account for integer precision above 2^53-1.",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "active_connections",
          "total_connections",
          "upload_bytes",
          "download_bytes",
          "history"
        ],
        "type": "object"
      },
      "SelectNodeRequest": {
        "additionalProperties": false,
        "properties": {
          "node_id": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "node_id"
        ],
        "type": "object"
      },
...
      "SelectorState": {
        "properties": {
          "track_all": {
            "type": "boolean",
            "description": "Read-only group track_all configuration. False by default: only the selected node is monitored. Configure the file and reload to change it."
          },
          "default_node_id": {
            "type": "string",
            "description": "Present only for an explicitly indexed selector(n). Plain selector has no configured default and cannot be reset."
          },
          "name": {
            "type": "string"
          },
          "node_id": {
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/SelectorNode"
            },
            "type": "array"
          },
          "overridden": {
            "type": "boolean",
            "description": "A saved preference exists, including when it is temporarily missing or ambiguous."
          },
          "saved_selection": {
            "$ref": "#/components/schemas/SavedSelectorSelection"
          }
        },
        "required": [
          "name",
          "node_id",
          "overridden",
          "track_all",
          "nodes"
        ],
        "type": "object"
      },
      "SelectorsResponse": {
        "properties": {
          "admin_enabled": {
            "type": "boolean"
          },
          "auth_mode": {
            "type": "string",
            "enum": ["api_key", "lan", "unix"],
            "description": "The request's administrative access mode: configured API key/session, verified direct LAN identity with no configured key, or filesystem-authorized Unix socket. LAN access is revalidated for every request and creates no browser session."
          },
          "selectors": {
            "items": {
              "$ref": "#/components/schemas/SelectorState"
            },
            "type": "array"
          }
        },
        "required": [
          "selectors",
          "admin_enabled",
          "auth_mode"
        ],
        "type": "object"
      },
      "ProbeRequest": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "outbound": { "type": "string", "minLength": 1, "description": "Name of an instantiated, checked outbound. Applies to selector and automatic policies, and checked direct-node targets." },
          "node_id": {
            "type": "string",
            "description": "Path ID within that outbound. Empty or omitted probes every path once."
          }
        },
        "required": ["outbound"]
      },
      "ProbeResponse": {
        "type": "object",
        "properties": {
          "outbound": { "type": "string" },
          "node_ids": { "type": "array", "items": { "type": "string" }, "description": "Accepted targets, including requests coalesced with queued/running checks. Acceptance is not a health result." }
        },
        "required": ["outbound", "node_ids"]
      },
...
      "StatusSnapshot": {
        "properties": {
          "direct_fallback_connections": {
            "description": "Process-lifetime count of established no-connectivity fallbacks to direct; preserved across reloads. Not a per-path or per-node counter.",
            "format": "int64",
            "type": "integer"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/GroupStatus"
            },
            "type": "array"
          },
          "last_reload_at": {
            "format": "date-time",
            "type": "string"
          },
          "plugins": {
            "items": {
              "$ref": "#/components/schemas/PluginInstanceStatus"
            },
            "type": "array"
          },
          "networks": {
            "description": "Network order: tcp4, tcp6, udp4, udp6.",
            "items": {
              "$ref": "#/components/schemas/PathStats"
            },
            "maxItems": 4,
            "minItems": 4,
            "type": "array"
          },
          "schema": {
            "const": 12,
            "format": "int64",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "type": "string"
          },
          "stats": {
            "$ref": "#/components/schemas/PathStats"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/TableUsage"
            },
            "type": "array"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "schema",
          "version",
          "started_at",
          "last_reload_at",
          "stats",
          "direct_fallback_connections",
          "networks",
          "tables",
          "groups"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "The configured global.api_key, sent as a Bearer credential.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dae_session",
        "description": "Signed seven-day browser session issued by PUT /api/session. HttpOnly, SameSite=Strict; scoped to /api and the listener host/port. Changing global.api_key invalidates it. An explicit Authorization header takes precedence."
      }
    }
  },
  "info": {
    "description": "Status schema 12. Direct fallback connections are a single process-lifetime counter outside path statistics. Node latency includes successful probes only; selection reports degradation and scoring independently. Domain tables report distinct retained domain/IP counts, kernel candidates and time-based GC; limit zero denotes unbounded userspace retention. TCP uses the router literal IP and configured api_port. The mode-0600 Unix socket /var/run/dae.sock provides local administration. Device self-service requires a direct TCP LAN connection. Mutations require X-Dae-API: 1. No query parameters; request bodies are limited to 1 KiB. Same-origin browser requests only; forwarding headers are not trusted. Successful operations return JSON, except certificate downloads and session changes (204). GET routes also accept HEAD. Counters are process-lifetime; poll with backoff on 503 during reload.",
    "title": "dae runtime API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/certificate": {
      "get": {
        "operationId": "getCertificate",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Certificate"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Read public CA identity"
      }
    },
    "/api/device": {
      "get": {
        "operationId": "getDevice",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceState"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Read the direct LAN caller's device state (TCP only)"
      }
    },
    "/api/device/status": {
      "get": {
        "operationId": "getDeviceStatus",
        "summary": "Read the direct LAN caller's identity and upstream runtime statistics",
        "responses": {
          "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeviceStatus" } } } },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/device/certificate-tests": {
      "post": {
        "operationId": "startCertificateTest",
        "summary": "Create a current-browser CA and transparent MITM challenge (direct LAN only)",
        "description": "Empty body. Does not change the device's MITM setting. GET trust_url and, when enabled, mitm_url from the browser with credentials omitted and redirects rejected. Dedicated HTTPS endpoints accept only the original page's Origin and return CertificateTestProof. The ordinary MITM-port TLS listener reports bypass; only the interception plugin reports mitm. These exact-origin diagnostics are the only cross-origin browser endpoints.",
        "parameters": [{ "name": "X-Dae-API", "in": "header", "required": true, "schema": { "type": "string", "const": "1" } }],
        "responses": {
          "201": { "description": "Challenge created, expires in one minute", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CertificateTest" } } } },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/device/certificate-tests/{id}": {
      "get": {
        "operationId": "getCertificateTest",
        "summary": "Read this device's challenge observations",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Current observation; 404 for expired, replaced or foreign challenges", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CertificateTest" } } } },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/selectors": {
      "get": {
        "description": "On TCP, a configured global.api_key requires a Bearer credential or valid browser session. When the key is empty, only a verified direct LAN connection is administrative, without credentials; failed LAN identification returns 403. A filesystem-authorized Unix socket connection is already administrative.",
        "operationId": "getSelectors",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectorsResponse"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          { "bearerAuth": [] },
          { "sessionCookie": [] },
          {}
        ],
        "summary": "List manual selectors (requires administration)"
      }
    },
    "/api/selectors/{group}": {
      "delete": {
        "description": "On TCP, a configured global.api_key requires a Bearer credential or valid browser session. When the key is empty, only a verified direct LAN connection is administrative, without credentials; failed LAN identification returns 403. A filesystem-authorized Unix socket connection is already administrative.",
        "operationId": "resetSelector",
        "parameters": [
          {
            "description": "URL-encoded name, including any slash.",
            "in": "path",
            "name": "group",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Dae-API",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectorState"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          },
          {}
        ],
        "summary": "Restore an explicitly configured selector(n) default; 409 when none exists"
      },
      "put": {
        "description": "On TCP, a configured global.api_key requires a Bearer credential or valid browser session. When the key is empty, only a verified direct LAN connection is administrative, without credentials; failed LAN identification returns 403. A filesystem-authorized Unix socket connection is already administrative.",
        "operationId": "selectNode",
        "parameters": [
          {
            "description": "URL-encoded name, including any slash.",
            "in": "path",
            "name": "group",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Dae-API",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectNodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectorState"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          },
          {}
        ],
        "summary": "Override a selector"
      }
    },
    "/api/plugins/{instance}/scripts/run": {
      "post": {
        "operationId": "triggerScript",
        "summary": "Trigger an existing cron or generic task",
        "description": "Requires administration. Returns 202 with an acceptance snapshot and per-task run number; execution belongs to the daemon and continues after the HTTP response. Inspect the plugin status for completion. Concurrent triggers for a waiting/running task return 409. Does not change its timer schedule. Unknown task/instance: 404; ambiguous name: 400; inactive worker: 503. No inline scripts, source URLs, argument overrides or automatic retries.",
        "parameters": [
          { "name": "instance", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "X-Dae-API", "in": "header", "required": true, "schema": { "type": "string", "const": "1" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScriptRunRequest" } } }
        },
        "responses": {
          "202": { "description": "Script execution accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScriptRunResponse" } } } },
          "default": { "$ref": "#/components/responses/Error" }
        },
        "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }, {}]
      }
    },
    "/api/probes": {
      "post": {
        "operationId": "probe",
        "summary": "Request configured outbound connectivity probes",
        "description": "Requires administration. Uses the outbound's configured DNS probe, existing concurrency bounds and timeouts. No arbitrary URL, network, or timeout overrides are accepted. node_id selects one path; omission or an empty value selects all paths. Returns 202 immediately with accepted IDs; repeated requests for queued/running work coalesce. Selector results are available through GET /api/selectors (checking/tested/checked_at/healthy/latency_ms); other outbounds publish their runtime health and latency through GET /api/status. This is not a persistent probe job and does not change selection or monitoring configuration. Unknown outbound: 404; invalid node/body: 400; unchecked builtin: 409.",
        "parameters": [
          { "name": "X-Dae-API", "in": "header", "required": true, "schema": { "type": "string", "const": "1" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProbeRequest" } } }
        },
        "responses": {
          "202": { "description": "Probe targets accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProbeResponse" } } } },
          "default": { "$ref": "#/components/responses/Error" }
        },
        "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }, {}]
      }
    },
    "/api/resources": {
      "get": {
        "operationId": "resourceRefreshStatus",
        "summary": "Read the latest resource refresh and next automatic check",
        "description": "Requires administration. Reports the latest API or automatic refresh. History is reset on daemon restart. Empty request body.",
        "responses": {
          "200": { "description": "Resource refresh status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResourceRefreshStatus" } } } },
          "default": { "$ref": "#/components/responses/Error" }
        },
        "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }, {}]
      }
    },
    "/api/resources/refresh": {
      "post": {
        "operationId": "refreshResources",
        "summary": "Refresh subscriptions, modules and dependencies",
        "description": "Requires administration and an empty request body. Uses the accepted configuration and attempts all downloads regardless of update intervals; does not read configuration edits from disk. Returns 202 immediately; check GET /api/resources for completion. Concurrent queued/running refreshes return 409. Refreshes and reloads publish serially. Invalid candidates leave the current runtime active; available validated snapshots can be used on download failure.",
        "parameters": [
          { "name": "X-Dae-API", "in": "header", "required": true, "schema": { "type": "string", "const": "1" } }
        ],
        "responses": {
          "202": { "description": "Resource refresh queued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResourceRefreshStatus" } } } },
          "default": { "$ref": "#/components/responses/Error" }
        },
        "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }, {}]
      }
    },
    "/api/session": {
      "put": {
        "operationId": "login",
        "summary": "Create a browser session",
        "description": "With a configured global.api_key, requires Authorization: Bearer <api_key>, including on Unix, and issues a signed seven-day session cookie without the API key. Missing/incorrect credentials return 401. Without a configured key, requires a verified direct LAN connection or filesystem-authorized Unix access, returns 204 and clears stale cookies without issuing a session; failed LAN identification returns 403. Empty request body.",
        "parameters": [
          {
            "name": "X-Dae-API",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "const": "1" }
          }
        ],
        "security": [ { "bearerAuth": [] }, {} ],
        "responses": {
          "204": {
            "description": "Access verified; no response body",
            "headers": {
              "Set-Cookie": {
                "description": "With a configured key: dae_session; Path=/api; HttpOnly; SameSite=Strict; Max-Age=604800. Without a key: expires any stale dae_session cookie instead.",
                "schema": { "type": "string" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "operationId": "logout",
        "summary": "Clear the browser session cookie",
        "description": "Empty request body. Clears the caller's cookie even if its session is expired. Does not require authentication.",
        "parameters": [
          {
            "name": "X-Dae-API",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "const": "1" }
          }
        ],
        "responses": {
          "204": {
            "description": "Cookie cleared; no response body",
            "headers": {
              "Set-Cookie": {
                "description": "Expires the dae_session cookie at Path=/api",
                "schema": { "type": "string" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/status": {
      "get": {
        "description": "On TCP, a configured global.api_key requires a Bearer credential or valid browser session. When the key is empty, only a verified direct LAN connection is administrative, without credentials; failed LAN identification returns 403. A filesystem-authorized Unix socket connection is already administrative.",
        "operationId": "getStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusSnapshot"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          },
          {}
        ],
        "summary": "Read the runtime snapshot"
      }
    },
    "/ca.cer": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/pkix-cert": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Certificate download"
          },
          "404": {
            "description": "Certificate unavailable"
          }
        },
        "summary": "Download public CA material; 404 when unavailable"
      }
    },
    "/ca.mobileconfig": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/x-apple-aspen-config": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Certificate download"
          },
          "404": {
            "description": "Certificate unavailable"
          }
        },
        "summary": "Download public CA material; 404 when unavailable"
      }
    },
    "/ca.pem": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/x-pem-file": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Certificate download"
          },
          "404": {
            "description": "Certificate unavailable"
          }
        },
        "summary": "Download public CA material; 404 when unavailable"
      }
    }
  },
  "servers": [
    {
      "description": "Replace with the router's actual literal IP and api_port",
      "url": "http://192.168.1.1:9080"
    }
  ]
}